listo el productsService
