+++ data/productsService.js
@@ -1,6 +1,6 @@
 const path = require('path');
 const fs = require('fs');
-const { viewCategory } = require('../controllers/productosController');
+const { viewCategory, viewDiscounts } = require('../controllers/productosController');
 const { create } = require('domain');
 
 const productsFilePath = path.join(__dirname, '../data/json-products.json');
@@ -56,17 +56,109 @@ const productsService = {
                 platform: req.body.platform,
                 stock: parseInt(req.body.stock)
             }
-            // console.log(youtubeId);
-            // console.log(newProduct);
+
             this.products.push(newProduct);
             let JSONproducts = JSON.stringify(this.products);
             fs.writeFileSync(productsFilePath, JSONproducts);
                 
         } catch (error) {
-            return 'error'
+            return error;
 
             // res.status(500).send(`No se seleccionaron las imágenes correspondientes. Seleccione las imágenes
             // que desea cargar e intente nuevamente.`);
+        }
+    },
+
+    viewEdit: function(id){
+        let productToEdit = this.products.find(product => product.id == id);
+        return {product : productToEdit};
+    },
+
+    viewDiscounts: function(){
+        let discounts = this.products.filter(product => product.discount > 0);
+        return {products: discounts};
+    },
+
+    edit: function(req){
+        let productID = req.params.id;
+        let productToEdit = this.products.find(product => product.id == productID);
+
+        // VERIFICAR SI SE CARGARON NUEVAS IMAGENES
+        let mainImage = "";
+        let moreImages = [];
+        let bannerImage = "";
+        if(!req.files['mainImage'] || req.files['mainImage'][0] == undefined){
+            mainImage = productToEdit.mainImage;
+        } else {
+            mainImage = req.files['mainImage'][0];
+        }
+
+        if(!req.files['moreImages'] || req.files['moreImages'][0] == undefined && 
+        req.files['moreImages'][1] == undefined && req.files['moreImages'][2] == undefined){
+            moreImages = productToEdit.moreImages;
+        } else {
+            moreImages = req.files['moreImages'];
+        }
+
+        if(!req.files['bannerImage'] || req.files['bannerImage'][0] == undefined){
+            bannerImage = productToEdit.bannerImage;
+        } else {
+            bannerImage = req.files['bannerImage'][0];
+        }
+
+        // OBTENER ID DE VIDEO DE YOUTUBE
+        function extractYouTubeId(url) {
+            const match = url.match(/[?&]v=([^?&]+)/);
+            return match ? match[1] : null;
+        }
+        const youtubeId = extractYouTubeId(req.body.video);
+
+		let updatedProduct = {
+			id: parseInt(productID), // Convert ID to integer
+			name: req.body.name,
+            price: parseFloat(req.body.price),
+            video: youtubeId,
+            description: req.body.description,
+            mainImage: mainImage,
+            moreImages: moreImages,
+            bannerImage: bannerImage,
+            category: req.body.category,
+            discount: parseFloat(req.body.discount),
+            finalPrice : parseFloat(req.body.price) - (parseFloat(req.body.price) * (parseFloat(req.body.discount) / 100)),
+            format: req.body.format,
+            platform: req.body.platform,
+            stock: req.body.stock
+		};
+	
+		// Find the index of the product in the products array
+		let productIndex = this.products.findIndex(product => product.id == productID);
+	
+		if (productIndex !== -1) { // Product found
+			// Update the product in the array
+			this.products[productIndex] = updatedProduct;
+	
+			// Convert the updated products array to JSON
+			let jsonProducts = JSON.stringify(this.products);
+	
+			// Write the updated JSON back to the file
+			fs.writeFileSync('./data/json-products.json', jsonProducts);
+		} else {
+			// Product not found
+			return "Product not found";
+		}
+    },
+
+    delete: function(id){
+        let productID = parseInt(id);
+        let productToDelete = this.products.find(product => product.id == productID);
+        let productIndex = this.products.findIndex(product => product.id == productID);
+        // console.log(productToDelete);
+        if (productIndex !== -1){
+            this.products.splice(productIndex, 1);
+            let jsonProducts = JSON.stringify(this.products);
+            fs.writeFileSync('./data/json-products.json', jsonProducts);
+        } else {
+            return 'producto no encontrado';
         }
     }
 }
